src/Services/PromptGen.API/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
redis==5.0.1
aioredis==2.0.1
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
openai==1.6.0
anthropic==0.8.0
numpy==1.24.3
jinja2==3.1.2
python-dotenv==1.0.0
slowapi==0.1.9
Pillow==10.1.0
